raised error messages if file does not exist to load

+++ TH Game/trunk/headers/Direct3DEngine.hpp
@@ -178,7 +178,7 @@ public:
 	void SetFog( float fognear, float fogfar, float fred, float fgreen, float fblue );
 	void SetPerspectiveClip( float zNear, float zFar )
 	{
-		D3DXMatrixPerspectiveFovLH( &ProjectionMatrix, D3DXToRadian( 45 ), 640.f/480.f, zNear, zFar );
+		D3DXMatrixPerspectiveFovLH( &ProjectionMatrix, (float)D3DXToRadian( 45.f ), 640.f/480.f, zNear, zFar );
 	}
 
 	//constructor
